+++ UnityCsprojNuget/Assets/UnityCsprojNuget.Editor/Application/FileResources.cs.meta
