+++ UnityCsprojNuget/Assets/UnityCsprojNuget.Editor/Application/IProjectCreator.cs
@@ -1,6 +1,6 @@
-﻿using UnityCsprojNuget.Editor.Bll.Entities;
+﻿using UnityCsprojNuget.Editor.Domain;
 
-namespace UnityCsprojNuget.Editor.Bll
+namespace UnityCsprojNuget.Editor.Application
 {
     public interface IProjectCreator
     {
