+++ UnityCsprojNuget/Assets/UnityCsprojNuget.Editor/Application/IProjectDiscoverer.cs.meta
