+++ UnityCsprojNuget/Assets/UnityCsprojNuget.Editor/Application/NamesPaths.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 
-namespace UnityCsprojNuget.Editor.Bll
+namespace UnityCsprojNuget.Editor.Application
 {
     public static class NamesPaths
     {
