+++ UnityCsprojNuget/Assets/UnityCsprojNuget.Editor/Application/ProjectDiscoverer.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using UnityCsprojNuget.Editor.Bll.Entities;
+using UnityCsprojNuget.Editor.Domain;
 
-namespace UnityCsprojNuget.Editor.Bll
+namespace UnityCsprojNuget.Editor.Application
 {
     public sealed class ProjectDiscoverer : IProjectDiscoverer
     {
