+++ UnityCsprojNuget/Assets/UnityCsprojNuget.Editor/Application/ProjectDiscoverer.cs.meta
