+++ UnityCsprojNuget/Assets/UnityCsprojNuget.Editor/Domain/ProjectDescriptor.cs.meta
