+++ UnityCsprojNuget/Assets/UnityCsprojNuget.Editor/SolutionPostProcessor.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-using UnityCsprojNuget.Editor.Bll;
+using UnityCsprojNuget.Editor.Application;
 using UnityEditor;
 
 namespace UnityCsprojNuget.Editor
